+++ docs/changelog.md
@@ -13,6 +13,7 @@ As features stabilize some brief notes about them will accumulate here.
 
 #### New
 * ssh client now supports `BindAddress`. Thanks to [@gpanders](https://github.com/gpanders)! [#1875](https://github.com/wez/wezterm/pull/1875)
+* [PaneInformation.domain_name](config/lua/PaneInformation.md) and [pane:get_domain_name()](config/lua/pane/get_domain_name.md) which return the name of the domain with which a pane is associated. [#1881](https://github.com/wez/wezterm/issues/1881)
 
 #### Fixed
 * Flush after replying to XTGETTCAP. [#1850](https://github.com/wez/wezterm/issues/1850)